requirements.txt
apache-airflow>=2.0.0
pytest>=6.2.0
pytest-mock>=3.6.0
requests>=2.25.0
